% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_growth_model.R
\docType{package}
\name{package_sgmodel}
\alias{package_sgmodel}
\alias{package_sgmodel-package}
\title{sgmodel: A package for computating the solutions to a generic stochastic growth model.}
\description{
The sgmodel package provides three important functions:
\code{sgmod}, \code{util} and \code{Markovmoments}.
}
\section{The \code{sgmodel} function}{

The \code{sgmodel} function solves a standard stochastic growth model using value function iteration. The stochastic component follows an autoregressive process of order one, and is discretized by a finite state Markov process.
}

\section{The \code{util} function}{

 It computes values for various uility functions encountered in economic theory.
}

\section{The \code{Markovmoments} function}{

It computes the four moments of a finite state Markov chain: expectation, variance, autocovariance and autocorrelation.
}

