% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_growth_model.R
\name{print_sgmod}
\alias{print_sgmod}
\title{print_sgmod}
\usage{
print_sgmod(x, ...)
}
\arguments{
\item{x}{A \code{sgmodel} object.}

\item{...}{Additional arguments.}
}
\value{
The function prints the call of the function, the \emph{Savings}, \emph{Consumption} and \emph{Capital grid} vectors  from \code{sgmodel}.
}
\description{
The function \code{print_sgmod} prints results of the \code{sgmodel} function.
}
\examples{
grid <- 200
utiltype <- "CRRA"
utilparam <- 4
A <- 1
depre <- 0.03
discount <- 0.95
prod <- 0.3
states <- 3
m <- 5
rho <- 0.2
sigma <- 0.02
model <- sgmodel(grid, utiltype, utilparam, A, depre, discount, prod, states, m, rho, sigma)
print_sgmod(model)
}
